requirements.txt
fastapi == 0.115.8
uvicorn == 0.34.0
mysql-connector-python == 9.2.0
